hello git are you fine

+++ app/Http/Controllers/ProductController.php
@@ -4,20 +4,36 @@ namespace App\Http\Controllers;
 
 use App\Product;
 use Illuminate\Http\Request;
+use Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Redirect;
 
 class ProductController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index');
+    }
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+
     public function index()
     {
+        $products = Product::orderBy('created_at', 'desc')->get();
+       // dd($posts); //dump and die
+        return view('products.index', ['products' => $products]);
+    }
+
+    public function manage()
+    {
+        //$this->authorize('create');
         $products = Product::get();
         $productsDeleted = Product::onlyTrashed()->get();
        // dd($posts); //dump and die
-        return view('products.index', ['products' => $products,
+        return view('products.manage', ['products' => $products,
         'productsDeleted' => $productsDeleted]);
     }
 
@@ -37,6 +53,7 @@ class ProductController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Product::class);
         return view('products.create');
     }
 
@@ -50,19 +67,21 @@ class ProductController extends Controller
     {
         $validatedData = $request->validate([
             'title' => ['required' , 'min:5', 'max:255'],
-            'detail' => ['required', 'max:500'],
-            'price' => ['required', 'numeric']            
-        ]);  
-        $request->validate([
+            'category' => 'required',
             'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-        ]); 
-        $imageName = time().'.'.$request->image->extension();  
+            'detail' => ['required', 'max:500'],
+            'price' => ['required', 'numeric']
+        ]);
+
+        $imageName = time().'.'.$request->image->extension();
         $request->image->move(public_path('img_product'), $imageName);
-        $product = new Product();        
+        $product = new Product();
         $product->title = $validatedData['title'];
         $product->image = $imageName;
+        $product->category_id = $validatedData['category'];
         $product->detail = $validatedData['detail'];
         $product->price = $validatedData['price'];
+        $product->user_id = Auth::id();
         $product->save();
 
         return redirect()->route('products.show', ['product' => $product]);
@@ -87,7 +106,10 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        //
+       /* if (Gate::allows('update-product', $product)) {
+            return redirect()->route('products.show', ['product' => $product->id]);
+        }*/
+        $this->authorize('update', $product);
         return view('products.edit', ['product' => $product]);
     }
 
@@ -108,13 +130,13 @@ class ProductController extends Controller
 
         $request->validate([
             'image' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-        ]); 
+        ]);
         if ($request->image == NULL) {
             $imageName = $product->image;
         } else {
-            $imageName = time().'.'.$request->image->extension();  
+            $imageName = time().'.'.$request->image->extension();
             $request->image->move(public_path('img_product'), $imageName);
-        }        
+        }
 
         $title = $validatedData['title'];
         $detail = $validatedData['detail'];
@@ -134,16 +156,17 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Product $product)
-    {        
-        
+    {
+        $this->authorize('delete', $product);
         $product->delete();
         $products = Product::get();
-        return redirect()->route('products.index' , ['products' => $products]);
+        //$productsDeleted = Product::onlyTrashed()->get();
+        return redirect('/manage');
     }
 
     public function restore(Product $product)
     {
-        $product->restore();        
+        $product->restore();
         return $this->index();
     }
 }
